+++ jksoa-mq/jksoa-mq-registry/src/main/kotlin/net/jkcode/jksoa/mq/registry/IMqDiscovery.kt
@@ -1,0 +1,24 @@
+package net.jkcode.jksoa.mq.registry
+
+interface IMqDiscovery {
+    /**
+     * 监听topic分配变化
+     *
+     * @param listener 监听器
+     */
+    fun subscribe(listener: IMqDiscoveryListener)
+
+    /**
+     * 取消topic分配变化
+     *
+     * @param listener 监听器
+     */
+    fun unsubscribe(listener: IMqDiscoveryListener)
+
+    /**
+     * 发现topic分配
+     *
+     * @return 服务地址
+     */
+    fun discover(): TopicAssignment
+}