+++ jksoa-mq/jksoa-mq-registry/src/main/kotlin/net/jkcode/jksoa/mq/registry/TopicAssignment.kt
@@ -1,4 +1,4 @@
-package net.jkcode.jksoa.mq.broker
+package net.jkcode.jksoa.mq.registry
 
 // topic分配情况: <topic, serverName>, serverName就是协议ip端口
 typealias TopicAssignment = MutableMap<String, String>